requirements.txt
setuptools>=65.5.1
wheel>=0.38.4
pip>=23.0.1
Flask>=2.2.3
python-dotenv>=1.0.0
pandas>=1.5.3
SQLAlchemy>=2.0.7
psycopg2-binary>=2.9.6
waitress>=2.1.2
configparser>=5.3.0
wrds>=3.1.2 